// web/src/gen/pb/api/v1/data.ts
export interface ListDashboardsResponse {
    /**
     * @generated from protobuf field: repeated testgrid.api.v1.DashboardResource dashboards = 1
     */
    dashboards: DashboardResource[];
}
// ...
export interface DashboardResource {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * @generated from protobuf field: string link = 2
     */
    link: string;
    /**
     * Dashboard group to which the dashboard belongs to.
     * Empty if dashboard doesn't belong to any group.
     *
     * @generated from protobuf field: string dashboard_group_name = 3
     */
    dashboardGroupName: string;
}
// ...
class ListDashboardsResponse$Type extends MessageType<ListDashboardsResponse> {
    constructor() {
        super("testgrid.api.v1.ListDashboardsResponse", [
            { no: 1, name: "dashboards", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => DashboardResource }
        ]);
    }
    create(value?: PartialMessage<ListDashboardsResponse>): ListDashboardsResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.dashboards = [];
        if (value !== undefined)
            reflectionMergePartial<ListDashboardsResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListDashboardsResponse): ListDashboardsResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated testgrid.api.v1.DashboardResource dashboards */ 1:
                    message.dashboards.push(DashboardResource.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListDashboardsResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated testgrid.api.v1.DashboardResource dashboards = 1; */
        for (let i = 0; i < message.dashboards.length; i++)
            DashboardResource.internalBinaryWrite(message.dashboards[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.ListDashboardsResponse
 */
export const ListDashboardsResponse = new ListDashboardsResponse$Type();
// ...
class DashboardResource$Type extends MessageType<DashboardResource> {
    constructor() {
        super("testgrid.api.v1.DashboardResource", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "link", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "dashboard_group_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<DashboardResource>): DashboardResource {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.link = "";
        message.dashboardGroupName = "";
        if (value !== undefined)
            reflectionMergePartial<DashboardResource>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DashboardResource): DashboardResource {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* string link */ 2:
                    message.link = reader.string();
                    break;
                case /* string dashboard_group_name */ 3:
                    message.dashboardGroupName = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: DashboardResource, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* string link = 2; */
        if (message.link !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.link);
        /* string dashboard_group_name = 3; */
        if (message.dashboardGroupName !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.dashboardGroupName);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.DashboardResource
 */
export const DashboardResource = new DashboardResource$Type();